+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/MapboxMapScope.kt
@@ -1,13 +1,12 @@
 package com.mapbox.maps.extension.compose
 
-import androidx.compose.foundation.layout.LayoutScopeMarker
 import androidx.compose.runtime.Immutable
 import com.mapbox.maps.MapboxExperimental
 
 /**
  * A MapboxMapScope provides a scope for the children of MapboxMap.
  */
-@LayoutScopeMarker
+@MapboxMapScopeMarker
 @Immutable
 @MapboxExperimental
 public object MapboxMapScope {